requirements.txt
# ouwibo-agent/requirements.txt

# --- Core AI ---
groq>=0.9.0,<2.0.0

# --- Web Framework ---
fastapi>=0.115.0,<2.0.0
uvicorn>=0.32.0,<1.0.0

# --- Rate Limiting ---
slowapi>=0.1.9,<1.0.0

# --- Database ---
SQLAlchemy>=2.0.0,<3.0.0

# --- Tools ---
ddgs>=7.0.0
googlesearch-python>=1.2.3
translate>=3.6.1
yfinance>=0.2.40

# --- Utilities ---
python-dotenv>=1.0.0,<2.0.0
tzdata>=2024.1
